refactor: remove barrel files, replace main.ts with marking-menu.ts
- Delete all barrel index.ts files (layout/, move/, navigation/, recognizer/)
- Delete main.ts and main.test.ts (replaced by marking-menu.ts)
- Update all imports to reference module files directly
- Point src/index.ts to marking-menu.js
