+++ src/index.ts
@@ -1,1 +1,1 @@
-export { default } from './main.js';
+export { default } from './marking-menu.js';
