--- src/layout/index.ts
@@ -1,4 +1,0 @@
-export { default as createMenuLayout } from './menu.js';
-export { default as createStrokeCanvas } from './stroke.js';
-export { default as connectLayout } from './connect.js';
-export { default as createGestureFeedback } from './gesture-feedback.js';
