--- src/move/index.ts
@@ -1,4 +1,0 @@
-export { default as watchDrags } from './linear-drag.js';
-export { default as longMoves } from './long-move.js';
-export { default as dwellings } from './dwelling.js';
-export { default as draw } from './draw.js';
