--- src/recognizer/index.ts
@@ -1,1 +1,0 @@
-export { default } from './recognize-mm-stroke.js';
